--- src/components/Cart/Cart.jsx
@@ -1,5 +1,0 @@
-function Cart() {
-  return <div>Cart</div>;
-}
-
-export default Cart;
