+++ src/pages/Carts.jsx
@@ -1,0 +1,5 @@
+function Carts() {
+  return <div>Carts</div>;
+}
+
+export default Carts;
