+++ src/pages/Main.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ProductListContainer from '../components/ProductList/ProductListContainer';
+
+function Main() {
+  return <ProductListContainer />;
+}
+
+export default Main;
